Rename routes and files
