+++ src/routes/ProtectedRoute.js
@@ -8,10 +8,8 @@ export const ProtectedRoute = ( { component: Component, ...rest } ) => {
   return ( <Route
     { ...rest }
     render={ props => {
-      return localStorage.getItem( 'loggedIn' ) !== 'true' ||
-      ( localStorage.getItem( 'loggedIn' ) === 'true' &&
-        localStorage.getItem( 'userRegistered' ) === 'true' ) ?
-        <Component { ...props }/> : <Redirect to={ '/signup' }/>;
+      return localStorage.getItem( 'loggedIn' ) !== 'true' ?
+        <Redirect to={ '/dashboard' }/> : <Component { ...props }/>;
     } }
   /> );
 };
