+++ src/routes/RegisteredAndLoggedInRoute.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Redirect, Route } from 'react-router';
+import { useStateValue } from '../hooks/useStateValue';
+
+export const RegisteredAndLoggedInRoute = ( { component: Component, ...rest } ) => {
+  const [ state ] = useStateValue();
+  
+  return ( <Route
+    { ...rest }
+    render={ props => {
+      return localStorage.getItem( 'loggedIn' ) !== 'true' ||
+      ( localStorage.getItem( 'loggedIn' ) === 'true' &&
+        localStorage.getItem( 'userRegistered' ) === 'true' ) ?
+        <Component { ...props }/> : <Redirect to={ '/signup' }/>;
+    } }
+  /> );
+};
+
+export default RegisteredAndLoggedInRoute;
