+++ src/routes/index.js
@@ -1,5 +1,5 @@
-export * from './CreateOrgRoute'
-export * from './LoginRoute'
-export * from './OrgDashboardRoute'
 export * from './ProtectedRoute'
+export * from './LoginRoute'
+export * from './OrganizationRoute'
+export * from './RegisteredAndLoggedInRoute'
 export * from './SignupRoute'
